requirements.txt
# API and AI
google-generativeai>=0.3.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.80.0

# Voice and audio
gTTS>=2.3.2
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Web automation and text extraction
selenium>=4.10.0
undetected-chromedriver>=3.5.0
pytesseract>=0.3.10
Pillow>=9.5.0

# Smart home integration
python-kasa>=0.5.0

# Networking
requests>=2.31.0

# Messaging
python-telegram-bot>=20.3